+++ src/MpvNet/MpvClient.cs
@@ -346,6 +346,9 @@ public class MpvClient
 
     public string GetPropertyString(string name)
     {
+        if (Handle == IntPtr.Zero)
+            return "";
+
         mpv_error err = mpv_get_property(Handle, GetUtf8Bytes(name),
             mpv_format.MPV_FORMAT_STRING, out IntPtr lpBuffer);
 
@@ -364,6 +367,12 @@ public class MpvClient
 
     public void SetPropertyString(string name, string value)
     {
+        if (Handle == IntPtr.Zero)
+        {
+            Terminal.WriteError($"error setting property: {name} = {value}");
+            return;
+        }
+
         byte[] bytes = GetUtf8Bytes(value);
         mpv_error err = mpv_set_property(Handle, GetUtf8Bytes(name), mpv_format.MPV_FORMAT_STRING, ref bytes);